requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
SQLAlchemy==2.0.23
PyJWT==2.8.0
Werkzeug==2.3.7
Jinja2==3.1.2
pytz==2023.3
pillow==10.1.0
pycryptodome==3.19.0
requests==2.31.0
python-dateutil==2.8.2
itsdangerous==2.1.2
click==8.1.7
waitress==2.1.2
gunicorn==21.2.0